backend-service/requirements.txt
fastapi
black
uvicorn[standard]
numpy
pandas
joblib
scikit-learn
tritonclient[all]
sqlalchemy
psycopg2-binary
binance